backend/backendapi/requirements.txt
Django==2.2
djangorestframework==3.13.1
django-cors-headers==3.13.0